# Raise various CSR limits to increase throughput