… go.sum
… vendor/bitbucket.org/ww/goautoneg/BUILD
… vendor/bitbucket.org/ww/goautoneg/Makefile
… vendor/bitbucket.org/ww/goautoneg/README.txt
… vendor/bitbucket.org/ww/goautoneg/autoneg.go
… vendor/golang.org/x/oauth2/README.md
… vendor/golang.org/x/oauth2/google/default.go
… vendor/golang.org/x/oauth2/google/google.go
… vendor/golang.org/x/oauth2/internal/token.go
… vendor/golang.org/x/oauth2/jwt/jwt.go
… vendor/golang.org/x/oauth2/oauth2.go
… vendor/golang.org/x/oauth2/token.go
… vendor/golang.org/x/sys/unix/README.md
… vendor/golang.org/x/sys/unix/mkall.sh
… vendor/golang.org/x/sys/unix/mkerrors.sh
… vendor/golang.org/x/sys/unix/mksyscall.go
… vendor/golang.org/x/sys/unix/mksyscall_solaris.go
… vendor/golang.org/x/sys/unix/mksyscall_solaris.pl
… vendor/golang.org/x/sys/unix/sockcmsg_unix.go
… vendor/golang.org/x/sys/unix/syscall_aix.go
… vendor/golang.org/x/sys/unix/syscall_darwin.go
… vendor/golang.org/x/sys/unix/syscall_linux.go
… vendor/golang.org/x/sys/unix/syscall_linux_arm.go
… vendor/golang.org/x/sys/unix/syscall_linux_arm64.go
… vendor/golang.org/x/sys/unix/syscall_linux_riscv64.go
… vendor/golang.org/x/sys/unix/syscall_unix.go
… vendor/golang.org/x/sys/unix/zerrors_linux_386.go
… vendor/golang.org/x/sys/unix/zerrors_linux_amd64.go
… vendor/golang.org/x/sys/unix/zerrors_linux_arm.go
… vendor/golang.org/x/sys/unix/zerrors_linux_arm64.go