- 5->20 worker goroutines for approving/signing controllers
    - 20->100 QPS limit talking to k8s API
    - 5->100 QPS limit talking to signing endpoint
    
    This gives ~2x speedup for CSR processing.